Explicitly cast address to string when looking up recipient
A recent database change made many of the "address" fields
integers that refer to recipient IDs. Since we also support
old database versions where the address refers to a phone
number (as a string), we explicitly have to cast the new
integer fields to ensure we don't end up creating fallback
recipients with "unknown" addresses due to type mismatch.