Update url regex and add url protocol if absent
There are a number of ways we can approach this. The best would be to
match Signal and implement the AUTOLINK_WEB_URL from the Android source
code (see: https://android.googlesource.com/platform/frameworks/base/+/master/core/java/android/util/Patterns.java#402)
I implemented this but decided to not include it because it will become a
whole licensing issue with the patterns being Apache licensed and a notice
file needing to be included, etc. Instead, I looked around for a good url
regex and found this one by John Gruber which is in the public domain and
does a decent job. We can consider adjusting it as we see fit.